Move Membership types into a separate file, exported from types.ts
