+++ spec/unit/room-member.spec.ts
@@ -21,13 +21,13 @@ import { RoomMember, RoomMemberEvent } from "../../src/models/room-member";
 import {
     createClient,
     EventType,
-    KnownMembership,
     MatrixClient,
     RoomState,
     UNSTABLE_MSC2666_MUTUAL_ROOMS,
     UNSTABLE_MSC2666_QUERY_MUTUAL_ROOMS,
     UNSTABLE_MSC2666_SHARED_ROOMS,
 } from "../../src";
+import { KnownMembership } from "../../src/@types/membership";
 
 describe("RoomMember", function () {
     const roomId = "!foo:bar";
