+++ spec/unit/scheduler.spec.ts
@@ -5,7 +5,8 @@ import { defer, IDeferred } from "../../src/utils";
 import { MatrixError } from "../../src/http-api";
 import { MatrixScheduler } from "../../src/scheduler";
 import * as utils from "../test-utils/test-utils";
-import { KnownMembership, MatrixEvent } from "../../src";
+import { MatrixEvent } from "../../src";
+import { KnownMembership } from "../../src/@types/membership";
 
 jest.useFakeTimers();
